Create GitUtil class #10
This class contains static functions to maniupulate the git repo.
